Level C partially completed
An entrepreneur principal now is able to:
- Edit his or her investment rounds, which includes creating, updating,
and deleting them. An investment round can be modified as long as it is
not saved in final mode. An investment round can be deleted as long as
no investor has applied for it, independently from whether it is in
draft or final mode.
- List the applications to his or her investment rounds grouped by
ticker or creation date and time.
- Accept or reject the applications to his or her investment rounds.
Accepted applications need not be justified, but rejected applications
must have a justification.

An authenticated principal now is able to:
- List investment rounds only when they are in final mode.

An investor principal now is able to:
- Apply for an investment round as long as it has been published.